fix logic login

+++ routes/articles.js
@@ -7,27 +7,32 @@ const Comment = require('../models/comment')
 
 
 router.get('/new', async(req, res) => {
+  let userId = req.session.userId;
   let user = await User.findById(req.session.userId)
   console.log(user,"user");
-  res.render('articles/new', { article: new Article(),user })
+  res.render('articles/new', { article: new Article(),user,userId })
 })
 
 router.get('/edit/:id', async (req, res) => {
-  
+  let userId = req.session.userId;
   let article = await Article.findById(req.params.id)
   let user = await User.findById(req.session.userId)
   
   let user_id = article.User
   if (user_id === req.session.userId){
-    res.render('articles/edit', { article, user })
+    res.render('articles/edit', { article, user,userId })
   }
-  else{
-    res.render('/dashboard')
-  }  
+  // else{
+  //   res.render('/dashboard')
+  // }  
 })
 
 router.get('/:slug', async (req, res) => {
+  let userId = req.session.userId;
   let article = await Article.findOne({ slug: req.params.slug });
+  console.log(article.User,"llkjh");
+  let markdown = await User.findById(article.User);
+  article.markdown = markdown
   let cmt = await Comment.find({ article: article._id.toString() });
     // let user_cmt = await User.find({_id:cmt.User})
     // console.log("cmmm",cmt);
@@ -39,7 +44,7 @@ router.get('/:slug', async (req, res) => {
     (article).comment_list = cmt;
   // console.log("p",req);
   if (article == null) res.redirect('/')
-  res.render('articles/show', { article, cmt })
+  res.render('articles/show', { article, cmt,userId })
 })
 
 router.post('/', async (req, res, next) => {
@@ -60,9 +65,15 @@ router.post('/:id', async (req, res, next) => {
 }, saveArticleAndRedirect('/edit'))
 
 router.post('/delete/:id', async (req, res) => {
-  await Article.findByIdAndDelete(req.params.id);
-  await Comment.deleteMany({article:req.params.id})
-  res.redirect('/dashboard')
+  let cur_user = req.session.userId;
+  let art = await Article.findById(req.params.id).select("User")
+  console.log(art,"arrrrj");
+  if(cur_user===art.User){
+    await Article.findByIdAndDelete(req.params.id);
+    await Comment.deleteMany({article:req.params.id})
+    res.redirect('/dashboard')
+  }
+  // else res.send("It is not your Article!!!")
 })
 
 router.post('/like/:id', async (req, res) => {
@@ -126,6 +137,7 @@ router.post('/like/:id', async (req, res) => {
 
 function saveArticleAndRedirect(path) {
   return async (req, res) => {
+    let userId = req.session.userId;
     // console.log("rbody",req.body.status);
     // console.log("resbody",req.body);
 
@@ -143,7 +155,7 @@ function saveArticleAndRedirect(path) {
       article = await article.save()
       res.redirect(`/articles/${article.slug}`)
     } catch (e) {
-      res.render(`articles/${path}`, { article: article })
+      res.render(`articles/${path}`, { article: article,userId })
     }
   }
 }
